+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.2] - 2026-07-08
+
+### Fixed
+
+- Firefox centering issues for icons
+- Safari print pages were cut off on top
+
+
 ## [0.1.1] - 2026-07-08
 
 ### Added